separate main bits out into files
turns out that importing stuff in react is pretty easy! The toolbar code
gets long quickly so it's not feasible to continue putting everything in
1 file
